refactor(core): migrate rate limiting and numeric types to improve precision
- Changed rate_limit type from f64 to u32 and semantics from milliseconds
  to requests per second
- Updated all exchange implementations to use new rate_limit type
- Migrated BidAsk and MarkPrice structs to use Decimal for financial
  calculations
- Replaced f64-based arithmetic with Decimal operations for better
  precision
- Updated tests to use decimal macros and assertions
- Modified spread and percentage calculations to use Decimal math
- Adjusted rate limiter configuration to match new semantic
- Updated documentation examples to reflect Decimal usage
- Fixed property tests to use exact equality instead of floating point
  comparisons

+++ ccxt-core/src/exchange.rs
@@ -273,8 +273,8 @@ pub trait Exchange: Send + Sync {
     }
 
     /// Returns the rate limit (requests per second)
-    fn rate_limit(&self) -> f64 {
-        10.0
+    fn rate_limit(&self) -> u32 {
+        10
     }
 
     // ==================== Market Data (Public API) ====================
